git-svn-id: svn://123.57.173.36/meijia-parent@1758 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-app/src/main/java/com/simi/action/app/xcloud/CompanyStaffController.java
@@ -400,7 +400,7 @@ public class CompanyStaffController extends BaseController {
 		if (companyId.equals(0L)) {
 			CompanyAdminSearchVo searchVo1 = new CompanyAdminSearchVo();
 			searchVo1.setUserId(userId);
-			searchVo1.setIsCreate(1);
+			searchVo1.setIsCreate((short) 1);
 			XcompanyAdmin xCompanyAdmin = null;
 			List<XcompanyAdmin> rs = xCompanyAdminService.selectBySearchVo(searchVo1);
